feat(auth-service): add auth require
User delete endpoint now requires authorization

+++ auth-service/source/service/user/user.py
@@ -1,3 +1,5 @@
+from datetime import datetime
+
 from fastapi import Depends
 
 from database.engine import get_session, AsyncSession
@@ -7,13 +9,19 @@ from database.queries.user import (
     get_user,
     delete_user, DeleteStatus,
     get_user_by_username,
+    user_from_token
 )
 from kernel.hashing import get_hasher, Hasher
 from kernel.tokens import get_token_logic, TokenLogic
+from scheme.tokens import TokenScheme
 
 from ..interfaces.user import IUserService, UserCreate, UserResult, UserAuth, Tokens
 from ..exceptions import UnexpectedError
-from .exceptions import CreateUsernameTaken, UserNotFound, WrongPassword
+from .exceptions import (
+    CreateUsernameTaken, UserNotFound, 
+    WrongPassword, WrongTokenData,
+    TokenNotActual, PermissionDenied
+)
 
 
 class UserService(IUserService):
@@ -23,6 +31,18 @@ class UserService(IUserService):
         self.__session: AsyncSession = session
         self.__hasher:  Hasher = hasher
         self.__token:   TokenLogic = token_logic
+    
+    async def __get_user_from_token(self, token: TokenScheme) -> UserORM:
+        user_result = await user_from_token(
+            self.__session,
+            id=token.id,
+            creation_data=datetime.fromtimestamp(token.created),
+        )
+
+        if (error := user_result.error()) is not None:
+            raise TokenNotActual(error)
+        
+        return user_result.result()
     
     async def create(self, user_data: UserCreate) -> UserResult:
         password_hash = await self.__hasher.hash(user_data.password)
@@ -49,7 +69,19 @@ class UserService(IUserService):
         
         return UserResult.model_validate(user)
 
-    async def delete(self, id: int) -> None:
+    async def delete(self, access_token: str, id: int) -> None:
+        validate_status = self.__token.validate(access_token)
+
+        if (error := validate_status.error()) is not None:
+            raise WrongTokenData(error)
+        
+        token_scheme = validate_status.result()
+        
+        user = await self.__get_user_from_token(token_scheme)
+
+        if user.id != id and not user.is_admin:
+            raise PermissionDenied()
+
         status = await delete_user(self.__session, id)
 
         if status == DeleteStatus.Ok:
